Update Main Page with Details
Updated the Main Page to contain an area where the details of a selected game will be displayed. Removed Details Page and tab bar from app since they are not needed.

Added a label that will start out in the new details area and will disappear when the first game is selected from the list of games available. Also locked the orientation for both Android and iOS to Portrait mode.