Started UI Testing
Added multiple xunit testing projects. Added one specifically for iOS and one that will be shared across any platform. This took some tutorials and messing with to understand why certain thing were set the way they were, but the tests now pass and can each run on an iPhone simulator producing a screenshot png in the bin folder.

Used these links to get to this point:
- https://devblogs.microsoft.com/dotnet/dotnet-maui-ui-testing-appium/
- https://vladislavantonyuk.github.io/articles/Testing-.NET-MAUI-Application-using-Appium-and-xUnit/
- https://xunit.net/docs/shared-context