Added View Model and updated tests
Added a view model for the main page which will handle the connection from the GUI to the business logic and back. Updated the main page to be started/updated by the local file storage and updated the tests to handle that as well. All tests are passing except the one that has not been implemented yet. Fixed path combining in multiple places.